Updating script.js without referencing a new block
