+++ blocks/embed/embed.css
@@ -1,4 +1,4 @@
-main .embed {
+/* main .embed {
     width: unset;
     text-align: center;
     max-width: 800px;
@@ -62,4 +62,4 @@ main .embed {
     border-left: 6px solid;
     top: 4px;
     left: 7px;
-  }
+  } */
